Moved Navigator to a new file
